Create WarWords.php
This is the Main php code to read the war words from a table and display them.
If a link is entered into the URL text field and the "Update" button is clicked, that link is read into $data and $data is searched for any words that exist in the WarWords table. 
When a word or phrase matches, it is surrounded by the hypertext bold characters <b> and </b> 
If the Red checkbox is checked the bold is converted to red-bold in CSS style script.