Create User Instructions
Instructions for using WarWords: 
John Reuwer: How can we give non-violence the appeal that war has?  Tony Jenkins: At the most basic level we need to challenge our own personal use of violent metaphors. Stephen Brivati: Virtually every word used in the ‘fight against cancer when we attack the illness, bombard it with radiation to kill etc’. is derived from … war. 
Most people don’t see violence in our language. (World Beyond War participants excluded.) One activity that I recommend is take any on-line article or page and bold all the war and violent words and phrases. Instead of doing it manually, this program does it automatically. 
Click on this link to open WarWords: https://www.modelsw.com/wbw/WarWords.php
Using WarWords In One Paragraph:
When you start WarWords, all the war and violence words and phrases are listed. Paste a link from any web page in the URL: [text box] and click the “Update” button (or Push Enter). Then you can see that web page with the War Words set bold in the text. Try the Red check box option. Skip to Email Below.
Using WarWords Copy, Paste for Dummies: 
 
When you start it, all the war and violence words and phrases are listed:
act of war, aim, allies, ambush, ammunition, annihilate, … 
…, zeal, zealot, zeppelin, zigzag, zone
On a different tab or window display a different web page.
For example, click on this link to display a new web page:
https://www.nytimes.com/2018/04/26/opinion/supreme-court-new-civil-war.html
Right click > Copy to highlight this “nytimes” link and Copy it.
 
Click the empty URL: [text box] in the WarWords tab or window. 
 

Right Click > Paste your “nytimes” copied link in the box after the word URL:
 
Click the “Update” Button (or push the Enter key)
Then scroll down through the article and you can see the words set bold in the text:

And the weapons aren’t cannon balls but rather the threatened withholding of federal money from “sanctuary” cities and states, the plan for the next census that would have the effect of shrinking liberal states’ representation in Congress, and the cap on tax deductions that will strike at residents of cities and states where high tax rates support decent public services.
If you click the Check box labeled “Red” at the right end of the URL: [text box  ] 
 
Then click the “Update” button (or push the Enter key).  
Now the words are set Red-Bold in the text:

And the weapons aren’t cannon balls but rather the threatened withholding of federal money from “sanctuary” cities and states, the plan for the next census that would have the effect of shrinking liberal states’ representation in Congress, and the cap on tax deductions that will strike at residents of cities and states where high tax rates support decent public services.
Email:
There is an email box at the bottom right of the WarWords screen:
 
Please email the words you think are missing or words that should be removed.
Click on the Email image and an Empty email will come up.

Copy or paste your changes into this email. Be sure to say:
“Remove: list of words” 
AND/OR
“Add: list of words.” View the phrases as a complete word. i.e. remove or add a complete phrase but don’t change part of the phrase by adding or removing a word within the phrase. 
AND/OR
Missing: a word that is listed and a link that you used that has that word but fails to bold.
All words are lower case. They are checked for uppercase and add ‘s’ for plurals so please don’t include case or plurals to the list unless the plural is different: knife, knives.
Peace (not on the list)